requirements.txt
asgiref==3.8.1
Django==5.1.1
mysqlclient==2.2.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
djangorestframework==3.15.2 
requests==2.32.3
python-dotenv==1.0.1